% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrviz.R
\name{diversity}
\alias{diversity}
\title{Diversity}
\usage{
diversity(x, which = c("shannon", "simpson", "invsimpson"))
}
\arguments{
\item{x}{a vector of counts}

\item{which}{the diversity index to calculate, one or more of
\code{"shannon"}, \code{"simpson"}, or \code{"invsimpson"}}
}
\description{
Calculate diversity indices.
}
\examples{
x <- rpois(50, 5)
diversity(x, 'shannon')
diversity(x)

}
