% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrviz.R
\name{tcrdata}
\alias{tcrdata}
\title{tcr data}
\usage{
tcrdata(data, seq = NULL, cnt = NULL, labels = NULL)
}
\arguments{
\item{data}{a data frame}

\item{seq, cnt}{column names or indices of \code{data} of sequences  and
counts, respectively}

\item{labels}{optional time point labels}
}
\value{
A matrix in the proper format for \code{\link{tcrviz}}.
}
\description{
Convenience function to format data for \code{tcrviz}.
}
\examples{
## alternative data format
dat <- tcrviz:::cbindx(
  time1 = letters[1:12], n1 = 1:12,
  space = NA,
  time2 = letters[2:10], n2 = 2:10,
  space = NA,
  time3 = letters[1:10], n3 = 1:10,
  space = NA,
  time4 = letters[6:15], n4 = 6:15
)
dat <- data.frame(dat)
dat[] <- lapply(dat, type.convert)

## data in tcrviz format
tcrdata(dat)
tcrdata(dat, labels = paste0('week', 1:4))

}
