% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcrviz.R
\name{tcrviz}
\alias{tcrviz}
\title{tcrviz}
\usage{
tcrviz(
  data,
  n.group = rep_len(1L, ncol(data)),
  group = rep_len("", length(n.group)),
  col.group = seq_along(group),
  col.gmain = col.group,
  labels = rep_len("", length(group)),
  show_bar = TRUE,
  show_pie = TRUE,
  show_text = TRUE,
  denom = seq.int(ncol(data)),
  show_counts = FALSE,
  n_min = NULL,
  inner_pie = TRUE,
  track_type = c("all", "detected", "none"),
  which = NULL,
  xlim = NULL,
  ylim = NULL,
  ylim_type = c("all", "detected", "post"),
  col.undetected = "grey90",
  cex.main = 2,
  heights = NULL,
  widths = NULL,
  space = 5,
  file = "",
  top = Inf,
  highlight = NULL,
  reset_par = TRUE,
  indices = c("shannon", "simpson", "invsimpson", "none")
)
}
\arguments{
\item{data}{an \code{n x t} count matrix with \code{n} unique sequences
for \code{t} time points}

\item{n.group, group}{grouping of time points: n per group and group label}

\item{col.group, col.gmain}{colors for each group and main group color}

\item{labels}{individual time point labels}

\item{show_bar, show_pie, show_text}{logicals; determine which sections will
be drawn: bar plots of counts, pie charts of proportions, and text
summaries, respectively}

\item{denom}{a vector (or list of vectors for each column of \code{data})
of column indices to get the denominator for number of unique sequences;
see examples}

\item{show_counts}{logical; if \code{TRUE}, counts are shown beside bars}

\item{n_min}{optional y-axis tick mark added at \code{n_min}, e.g., for
cutting singletons from figure}

\item{inner_pie}{logical; if \code{TRUE}, an inner pie chart added inside
main; the main pie is the proportion of each originating time point, and
the inner pie also includes undetected}

\item{track_type}{type of tracking arrows to draw between bars; one of
\code{"all"} (default, detected and undetected), \code{"detected"} (for
only detected), or \code{"none"} to suppress all arrows}

\item{which}{integer index of the column to highlight (i.e., all other
columns will be semi-transparent); use \code{0} to highlight all}

\item{xlim, ylim}{x- and y-axis limits}

\item{ylim_type}{extent of ylim shown (\code{"all"} - all unique sequences,
\code{"detected"} - max of number detected, \code{"post"} - max of number
detected after post) -- \code{all > detected > post}; ignored if
\code{ylim} is given}

\item{col.undetected}{color of undetected sequences (used in pie charts)}

\item{cex.main}{size of main labels}

\item{heights, widths}{see \code{\link{layout}}}

\item{space}{number of lines between top of bars from rest of plot}

\item{file}{output file name to write table; if none, no table is written}

\item{top}{for table output, only report \code{top} unique sequences by
frequency (use \code{Inf} for all)}

\item{highlight}{index of timepoint(s) to highlight (use 0 for all); e.g.,
if \code{highlight = 2}, the second column color is fully opaque with
transparency applied to other groups; see examples}

\item{reset_par}{logical; if \code{TRUE} (default), graphical parameters
are reset to state before plot call; use \code{FALSE} to continue adding
to figure}

\item{indices}{diversity indices to add to table output (all selected are
calculated for each column of \code{data}); one or more of \code{"shannon"},
\code{"simpson"}, or \code{"invsimpson"}; use \code{"none"} to exclude}
}
\description{
TCR tracking visualizations.
}
\examples{
## generate example data (matrix must have class "tcr" and dimnames)
set.seed(1)
l <- 5:1 / 10
dat <- structure(
  sapply(l, function(x)
    sample(sort(rnbinom(50, 1, x)), prob = seq.int(50))),
  .Dimnames = list(1:50, letters[1:5])
)

tcrviz(dat)
tcrviz(dat, highlight = 0) ## same as above
tcrviz(dat, highlight = 2) ## only second column is fully opaque

\dontrun{
## all opaque then highlight each
op <- par(ask = TRUE)
tcrviz(dat, highlight = 0:ncol(dat))
par(op)
}

tcrviz(
  dat, n.group = c(1, 2, 2), inner_pie = TRUE, n_min = 1,
  group = c('first', 'second', 'third'),
  # denom = list(1:5, 2:5, 2:5, 2:5, 2:5), ## cols 2-5 independent of 1
  col.group = c('grey', 'tomato', 'orange'),
  ## use different xlim for each panel
  xlim = lapply(1:5, function(x) c(0, max(dat[, x])))
)

}
\seealso{
\code{\link{diversity}}
}
